Revert "[DSCP-126] Use Word32 for nonce everywhere"
This reverts commit 4e8eef31c32ce85cc1b948a9c8bedb8f75bdd614.

+++ witness/src/Dscp/Witness/Web/Logic.hs
@@ -72,7 +72,7 @@ toAccountInfo account txs = do
     pure AccountInfo
         { aiBalances = Coin . fromIntegral . aBalance <$> account
         , aiCurrentNonce = nonce
-        , aiTransactionCount = fromIntegral nonce
+        , aiTransactionCount = nonce
         , aiTransactions = transactions
         }
   where
